+++ src/cplx-user-settings/types/cplx-user-settings.types.ts
@@ -47,7 +47,6 @@ export const cplxUserSettingsSchema = z.object({
     }),
   }),
   customTheme: z.object({
-    slogan: z.string().optional(),
     uiFont: z.string().optional(),
     monoFont: z.string().optional(),
     accentColor: z.string().optional(),